Added redux to App container
