+++ app/containers/App/selectors.js
@@ -1,3 +1,21 @@
+/*
+ * The global state selectors
+ */
+
+import { createSelector } from 'reselect';
+
+const selectGlobal = (state) => state.get('global');
+
+const makeSelectUser = () => createSelector(
+  selectGlobal,
+  (globalState) => globalState.get('user')
+);
+
+const makeSelectOnboardingValidation = () => createSelector(
+  selectGlobal,
+  (onboardingState) => onboardingState.get('validation')
+);
+
 // makeSelectLocationState expects a plain JS object for the routing state
 const makeSelectLocationState = () => {
   let prevRoutingState;
@@ -16,5 +34,8 @@ const makeSelectLocationState = () => {
 };
 
 export {
+  selectGlobal,
+  makeSelectUser,
   makeSelectLocationState,
+  makeSelectOnboardingValidation,
 };
