+++ app/containers/Onboarding/constants.js
@@ -1,10 +1,3 @@
 export const ONBOARDING_PREV_STAGE = 'venu/Onboarding/ONBOARDING_PREV_STAGE';
 export const ONBOARDING_NEXT_STAGE = 'venu/Onboarding/ONBOARDING_NEXT_STAGE';
 export const ONBOARDING_START_VENU = 'venu/Onboarding/ONBOARDING_START_VENU';
-export const CHANGE_USER_EMAIL = 'venu/Onboarding/CHANGE_USER_EMAIL';
-export const CREATE_USER_ACCOUNT = 'venu/Onboarding/CREATE_USER_ACCOUNT';
-export const CHANGE_USER_DISPLAYNAME = 'venu/Onboarding/CHANGE_USER_DISPLAYNAME';
-export const ERROR_USER_ACCOUNT_CREATION = 'venu/Onboarding/ERROR_USER_ACCOUNT_CREATION';
-export const SETUP_GEOLOCATION = 'venu/Onboarding/SETUP_GEOLOCATION';
-export const CHANGE_PARKING_LOCATION = 'venu/Onboarding/CHANGE_PARKING_LOCATION';
-export const CHANGE_USER_INTERESTS = 'venu/Onboarding/CHANGE_USER_INTERESTS';
