+++ app/reducers.js
@@ -7,8 +7,10 @@ import { combineReducers } from 'redux-immutable';
 import { fromJS } from 'immutable';
 import { LOCATION_CHANGE } from 'react-router-redux';
 
-import languageProviderReducer from 'containers/LanguageProvider/reducer';
+import globalReducer from 'containers/App/reducer';
 import onboardingReducer from 'containers/Onboarding/reducer';
+import languageProviderReducer from 'containers/LanguageProvider/reducer';
+
 /*
  * routeReducer
  *
@@ -43,6 +45,7 @@ function routeReducer(state = routeInitialState, action) {
 export default function createReducer(asyncReducers) {
   return combineReducers({
     route: routeReducer,
+    global: globalReducer,
     language: languageProviderReducer,
     onboarding: onboardingReducer,
     ...asyncReducers,
